Warning - potentially breaking changes on some BBs (see commit message)
May need to revert to previous commit, as the current script fails to
pre-compile on the newly flashed BBs. Currently, opening a julia prompt on the
BB and running

    include("/home/debian/juliapackages/LabConnections/src/BeagleBone/startup/startup.jl")

runs. However, if running this from the julia.server sript upon booting the BB,
the package 'LabConnections' cannot be found. If disbling the service,
pushing the path in julia allows commands such as

    using LabConnections

run, but the tests (which ran on my previous BB installation) no longer pass.
These have not been changed since the most recent commit.